+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to semantic versioning.
 - Add `unsafeExpect` and `unsafeExpectLeft` to `IOEither`.
 - Add `whenM` and `unlessM` to `Monad`.
 - Add `unsafeExpect` to `Option`.
+- Add `asksTask` to `ReaderTask`.
 - Add `unsafeExpect` and `unsafeExpectLeft` to `TaskEither`.
 - Add `unsafeExpect` to `TaskOption`.
 - Remove `flip`. [fp-ts/function::flip](https://gcanti.github.io/fp-ts/modules/function.ts.html#flip) is now curried.